Reworked the format of batch parameter files.
Now there is a mandatory parameter, $label$, that must be defined. You should
have it include the parameters that are varying from run to run.

The "short tokens" are removed. The format for a parameter line has
changed from:
short_token;$token$;value1;value2;value3...
to
$token$;value1;value2;value3...

To have this same effect, set the label line to:
$label$;short_token1$token1$short_token2$token2$short_token3$token3$


git-svn-id: svn+ssh://eggletonlinux.meng.umbc.edu/scratch/linux1/svn_repo/trunk/immersed_boundary/cell@538 3eb578c0-666c-0410-8428-3510748880ad